Add functionality to save posts in an array Add cSpell words and update homeStartingContent, aboutContent, and contactContent
This commit introduces the ability to save each new post in a global array. When a POST request is made to the '/compose' route, a new post object is created from the request body and added to the 'posts' array. This allows us to keep track of all the posts that have been made. The entire 'posts' array is then logged to the console for debugging purposes.