Add post creation functionality
This commit adds a new route handler for '/compose' in app.js. When a POST request is made to '/compose', a new post is created with the title and content from the request body. The post is then added to the 'posts' array and the client is redirected to the home page.